Rename field_value.ts to object_value.ts
After the Field Value migration, there is no FieldValue type anymore. This renames the corresponding file.
