+++ packages/firestore/src/model/document.ts
@@ -21,7 +21,7 @@ import { SnapshotVersion } from '../core/snapshot_version';
 import { fail } from '../util/assert';
 
 import { DocumentKey } from './document_key';
-import { ObjectValue } from './field_value';
+import { ObjectValue } from './object_value';
 import { FieldPath } from './path';
 import { valueCompare } from './values';
