+++ packages/firestore/src/remote/serializer.ts
@@ -34,7 +34,7 @@ import { TargetId } from '../core/types';
 import { TargetData, TargetPurpose } from '../local/target_data';
 import { Document, MaybeDocument, NoDocument } from '../model/document';
 import { DocumentKey } from '../model/document_key';
-import { ObjectValue } from '../model/field_value';
+import { ObjectValue } from '../model/object_value';
 import {
   DeleteMutation,
   FieldMask,
