+++ packages/firestore/test/util/api_helpers.ts
@@ -36,7 +36,7 @@ import {
 import { DocumentKeySet } from '../../src/model/collections';
 import { Document } from '../../src/model/document';
 import { DocumentSet } from '../../src/model/document_set';
-import { JsonObject } from '../../src/model/field_value';
+import { JsonObject } from '../../src/model/object_value';
 import { doc, key, path as pathFrom } from './helpers';
 import { Provider, ComponentContainer } from '@firebase/component';
